Move human_time and get_local_ip to util.py
